Moved embed to their own project to keep textcat lightweight
Bonus: Added support for Japan
