+++ .gitignore
@@ -2,6 +2,7 @@
 # Generated by Cargo
 # will have compiled files and executables
 /target/
+embed/target/
 
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
