+++ Cargo.toml
@@ -13,4 +13,3 @@ serde_json = "1.0"
 glob = "0.3.0"
 serde_bytes = "0.11"
 unicode-segmentation = "1.7.1"
-tera = "1"
