+++ embed/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "textcat-embed"
+version = "0.1.0"
+authors = ["Cesar Rodas <cesar@rodasm.com.py>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+textcat = {path="../"}
+tera = "1"
