+++ embed/samples/arabic.sample
