+++ embed/samples/catalan.sample
