+++ embed/samples/chinese.sample
