+++ embed/samples/danish.sample
