+++ embed/samples/english.sample
