+++ embed/samples/esperanto.sample
