+++ embed/samples/euskara.sample
