+++ embed/samples/french.sample
