+++ embed/samples/german.sample
