+++ embed/samples/hebrew.sample
