+++ embed/samples/irish.sample
