+++ embed/samples/japanese.sample
@@ -1,0 +1,19 @@
+　福岡県篠栗（ささぐり）町で昨年4月、5歳の男児が十分な食事を与えられず餓死した事件で、福岡地検は23日、母親の碇（いかり）利恵（39）と知人の赤堀恵美子（48）の両容疑者=ともに同町尾仲=を保護責任者遺棄
+
+　起訴状によると、赤堀容疑者は碇容疑者の一家に自らが提供する食事のみで質素な生活をするよう言い含め、生活全般を実質的に支配。その上で両容疑者は共謀し、2019年8月ごろから碇容疑者の三男翔士郎（しょうじろう）ちゃんの食事の量や回数を減らし、遅くとも昨年3月下旬ごろには重度の低栄養状態にさせ、碇容疑者が保護すべき責任があったのに十分な食事を与えず放置して昨年4月18日午後10時ごろ餓死させたとされる。
+
+　福岡県警は、赤堀容疑者が「他の保護者があなたの悪口を言っている」などとうそで不安にさせる一方、架空のトラブルを捏造（ねつぞう）しては解決したように装うことで信頼させ、指示通り動くようマインドコントロールしていったとみている。捜査関係者によると、赤堀容疑者の指示で碇一家の食事は極度に切り詰められ、翔士郎ちゃんは昨年3月には半月以上水だけしか与えられず、死亡時の体重は1歳6カ月の検診時と同じ約10キロだったという。
+
+　地検は、赤堀容疑者が保護責任者かどうかにかかわらず、碇容疑者に翔士郎ちゃんを餓死させるよう仕向ける言動があったとみて、碇容疑者の共謀共同正犯として同等の責任が問えると判断した。
+
+
+
+英アストラゼネカは22日、新型コロナウイルスのワクチンに「症状を防ぐ効果が79%あり、血栓を含む重大な副反応もなかった」とする最終段階の臨床試験（治験）の暫定結果を公表した。これに対し、米国立アレルギー・感染症研究所（NIAID）は23日、「治験の有効性に関するデータが不完全だった可能性がある」との異例の声明を発表。速やかにデータを検証するよう同社に求めた。
+
+　同社は、米国などで約3万2千人を対象にした治験の結果を公表。参加者の3分の2がワクチン、残りは偽薬を使い、効果や安全性を調べたところ、ワクチンが発症を79%減らし、重症化は100%防いだとした。また、欧州などで報告されて接種が一時中断した血栓の発生についても、治験では接種を受けた2万1千人余に血栓のリスク増加はなかったとした。
+
+　ところが、発表後にNIAIDが、「データに古い情報が含まれていた可能性がある」などとする声明を発表した。独立した立場にある効果安全性評価委員会から、治験データへの懸念を伝えられたという。治験はNIAIDなど米政府機関が予算支援していた。
+
+　同社のワクチンは英国や欧州、豪州、韓国などで使用許可が出ており、米食品医薬品局（FDA）にも近く緊急時使用の許可を申請するとしていた。日本も1億2千万回分を調達予定で、厚生労働省が承認審査をしている。同社は23日、「発表は2月17日までのデータを使った中間解析だった。評価委員会と連携し、最新の有効性を含む解析結果を共有する予定だ」との見解を出した。（ワシントン=香取啓介）
+
+
