+++ embed/samples/latin.sample
