+++ embed/samples/norwegian.sample
