+++ embed/samples/persian.sample
