+++ embed/samples/romanian.sample
