+++ embed/samples/swedish.sample
