+++ src/default.rs
@@ -1,4 +1,8 @@
-
+/// This file is automatically generated by `embed`. Any changes made here will be
+/// lost forever.
+///
+/// Take a look at the `embed` directory for instructions on how to add improvements
+/// or how to add a new language
 use crate::storage::FileContent;
 use std::str::FromStr;
 
@@ -20,6 +24,7 @@ pub enum Language {
     Hebrew,
     Irish,
     Italian,
+    Japanese,
     Ladino,
     Latin,
     Norwegian,
@@ -53,6 +58,7 @@ impl FromStr for Language {
             "hebrew" => Ok(Self::Hebrew),
             "irish" => Ok(Self::Irish),
             "italian" => Ok(Self::Italian),
+            "japanese" => Ok(Self::Japanese),
             "ladino" => Ok(Self::Ladino),
             "latin" => Ok(Self::Latin),
             "norwegian" => Ok(Self::Norwegian),
@@ -71,6 +77,7 @@ pub struct TextCat {
     built_in: FileContent,
 }
 
+#[allow(clippy::new_without_default)]
 impl TextCat {
     pub fn new() -> Self {
         TextCat {
@@ -82,6 +89,19 @@ impl TextCat {
         self.built_in
             .get_category(sample)
             .map(|r| Language::from_str(r.as_str()).unwrap())
+    }
+
+    pub fn get_languages(
+        &self,
+        sample: &str,
+    ) -> Option<Vec<(Language, u64)>> {
+        Some(
+            self.built_in
+                .get_categories(sample)?
+                .iter()
+                .map(|r| (Language::from_str(&r.0).unwrap(), r.1))
+                .collect(),
+        )
     }
 
     pub fn get_embed_languages() -> FileContent {
@@ -2194,6 +2214,411 @@ impl TextCat {
                 ],
             ),
             (
+                "japanese",
+                vec![
+                    "の_",
+                    "_の_",
+                    "_の",
+                    "た_",
+                    "_た_",
+                    "_た",
+                    "を_",
+                    "_を_",
+                    "_を",
+                    "に_",
+                    "_に_",
+                    "_に",
+                    "と_",
+                    "し_",
+                    "_と_",
+                    "_と",
+                    "_し_",
+                    "_し",
+                    "は_",
+                    "が_",
+                    "_は_",
+                    "_は",
+                    "_が_",
+                    "_が",
+                    "る_",
+                    "_る_",
+                    "_る",
+                    "者_",
+                    "_者_",
+                    "_者",
+                    "い_",
+                    "_い_",
+                    "_い",
+                    "て_",
+                    "_て_",
+                    "_て",
+                    "な_",
+                    "で_",
+                    "う_",
+                    "_な_",
+                    "_な",
+                    "_で_",
+                    "_で",
+                    "_う_",
+                    "_う",
+                    "疑_者_",
+                    "疑_者",
+                    "疑_",
+                    "容_疑_者",
+                    "容_疑_",
+                    "容_疑",
+                    "容_",
+                    "し_た_",
+                    "し_た",
+                    "か_",
+                    "_疑_者_",
+                    "_疑_者",
+                    "_疑_",
+                    "_疑",
+                    "_容_疑_",
+                    "_容_疑",
+                    "_容_",
+                    "_容",
+                    "_し_た_",
+                    "_し_た",
+                    "_か_",
+                    "_か",
+                    "ら_",
+                    "っ_",
+                    "_ら_",
+                    "_ら",
+                    "_っ_",
+                    "_っ",
+                    "_2",
+                    "っ_た_",
+                    "っ_た",
+                    "た_と_",
+                    "た_と",
+                    "す_",
+                    "さ_",
+                    "_っ_た_",
+                    "_っ_た",
+                    "_た_と_",
+                    "_た_と",
+                    "_す_",
+                    "_す",
+                    "_さ_",
+                    "_さ",
+                    "月_",
+                    "同_",
+                    "れ_",
+                    "す_る_",
+                    "す_る",
+                    "_月_",
+                    "_月",
+                    "_同_",
+                    "_同",
+                    "_れ_",
+                    "_れ",
+                    "_す_る_",
+                    "_す_る",
+                    "_1",
+                    "験_",
+                    "碇_",
+                    "日_",
+                    "り_",
+                    "よ_",
+                    "ど_",
+                    "て_い_",
+                    "て_い",
+                    "し_て_",
+                    "し_て",
+                    "あ_",
+                    "_験_",
+                    "_験",
+                    "_碇_",
+                    "_碇",
+                    "_日_",
+                    "_日",
+                    "_り_",
+                    "_り",
+                    "_よ_",
+                    "_よ",
+                    "_ど_",
+                    "_ど",
+                    "_て_い_",
+                    "_て_い",
+                    "_し_て_",
+                    "_し_て",
+                    "_あ_",
+                    "_あ",
+                    "3_",
+                    "食_",
+                    "表_",
+                    "発_",
+                    "治_験_",
+                    "治_験",
+                    "治_",
+                    "果_",
+                    "性_",
+                    "事_",
+                    "ろ_",
+                    "も_",
+                    "な_ど_",
+                    "な_ど",
+                    "だ_",
+                    "_食_",
+                    "_食",
+                    "_表_",
+                    "_表",
+                    "_発_",
+                    "_発",
+                    "_治_験_",
+                    "_治_験",
+                    "_治_",
+                    "_治",
+                    "_果_",
+                    "_果",
+                    "_性_",
+                    "_性",
+                    "_事_",
+                    "_事",
+                    "_ろ_",
+                    "_ろ",
+                    "_も_",
+                    "_も",
+                    "_な_ど_",
+                    "_な_ど",
+                    "_だ_",
+                    "_だ",
+                    "2_",
+                    "食_事_",
+                    "食_事",
+                    "赤_堀_",
+                    "赤_堀",
+                    "赤_",
+                    "者_の_",
+                    "者_の",
+                    "碇_容_疑",
+                    "碇_容_",
+                    "碇_容",
+                    "時_",
+                    "年_",
+                    "堀_",
+                    "効_",
+                    "ータ_",
+                    "ータ",
+                    "ン_",
+                    "データ_",
+                    "データ",
+                    "デー",
+                    "タ_",
+                    "る_と_",
+                    "る_と",
+                    "よ_う_",
+                    "よ_う",
+                    "は_2",
+                    "っ_た_と",
+                    "せ_",
+                    "さ_せ_",
+                    "さ_せ",
+                    "が_あ_",
+                    "が_あ",
+                    "え_",
+                    "_食_事_",
+                    "_食_事",
+                    "_赤_堀_",
+                    "_赤_堀",
+                    "_赤_",
+                    "_赤",
+                    "_者_の_",
+                    "_者_の",
+                    "_碇_容_",
+                    "_碇_容",
+                    "_時_",
+                    "_時",
+                    "_年_",
+                    "_年",
+                    "_堀_",
+                    "_堀",
+                    "_効_",
+                    "_効",
+                    "_ワ",
+                    "_データ_",
+                    "_データ",
+                    "_デー",
+                    "_デ",
+                    "_る_と_",
+                    "_る_と",
+                    "_よ_う_",
+                    "_よ_う",
+                    "_は_2",
+                    "_せ_",
+                    "_せ",
+                    "_さ_せ_",
+                    "_さ_せ",
+                    "_が_あ_",
+                    "_が_あ",
+                    "_え_",
+                    "_え",
+                    "_3",
+                    "_2_",
+                    "重_",
+                    "赤_堀_容",
+                    "責_任_",
+                    "責_任",
+                    "責_",
+                    "護_",
+                    "解_",
+                    "者_が_",
+                    "者_が",
+                    "米_",
+                    "社_",
+                    "発_表_",
+                    "発_表",
+                    "症_",
+                    "疑_者_の",
+                    "生_",
+                    "死_",
+                    "検_",
+                    "昨_年_",
+                    "昨_年",
+                    "昨_",
+                    "堀_容_疑",
+                    "堀_容_",
+                    "堀_容",
+                    "国_",
+                    "含_",
+                    "同_社_",
+                    "同_社",
+                    "可_",
+                    "分_",
+                    "共_",
+                    "全_",
+                    "保_護_",
+                    "保_護",
+                    "保_",
+                    "使_",
+                    "任_",
+                    "一_",
+                    "ワクチン_",
+                    "ワクチン",
+                    "ワクチ",
+                    "ワク",
+                    "チン_",
+                    "チン",
+                    "クチン_",
+                    "クチン",
+                    "クチ",
+                    "ら_れ_",
+                    "ら_れ",
+                    "や_",
+                    "と_し_",
+                    "と_し",
+                    "ず_",
+                    "し_て_い",
+                    "こ_",
+                    "_重_",
+                    "_重",
+                    "_責_任_",
+                    "_責_任",
+                    "_責_",
+                    "_責",
+                    "_護_",
+                    "_護",
+                    "_解_",
+                    "_解",
+                    "_者_が_",
+                    "_者_が",
+                    "_米_",
+                    "_米",
+                    "_社_",
+                    "_社",
+                    "_発_表_",
+                    "_発_表",
+                    "_症_",
+                    "_症",
+                    "_生_",
+                    "_生",
+                    "_死_",
+                    "_死",
+                    "_検_",
+                    "_検",
+                    "_昨_年_",
+                    "_昨_年",
+                    "_昨_",
+                    "_昨",
+                    "_堀_容_",
+                    "_堀_容",
+                    "_国_",
+                    "_国",
+                    "_含_",
+                    "_含",
+                    "_同_社_",
+                    "_同_社",
+                    "_可_",
+                    "_可",
+                    "_分_",
+                    "_分",
+                    "_共_",
+                    "_共",
+                    "_全_",
+                    "_全",
+                    "_保_護_",
+                    "_保_護",
+                    "_保_",
+                    "_保",
+                    "_使_",
+                    "_使",
+                    "_任_",
+                    "_任",
+                    "_一_",
+                    "_一",
+                    "_ワクチン",
+                    "_ワクチ",
+                    "_ワク",
+                    "_ら_れ_",
+                    "_ら_れ",
+                    "_や_",
+                    "_や",
+                    "_と_し_",
+                    "_と_し",
+                    "_ず_",
+                    "_ず",
+                    "_こ_",
+                    "_こ",
+                    "_3_",
+                    "9_",
+                    "験_の_",
+                    "験_の",
+                    "餓_死_",
+                    "餓_死",
+                    "餓_",
+                    "関_",
+                    "郎_",
+                    "言_",
+                    "血_栓_",
+                    "血_栓",
+                    "血_",
+                    "翔_士_郎",
+                    "翔_士_",
+                    "翔_士",
+                    "翔_",
+                    "結_果_を",
+                    "結_果_",
+                    "結_果",
+                    "結_",
+                    "立_",
+                    "福_岡_",
+                    "福_岡",
+                    "福_",
+                    "疑_者_が",
+                    "状_",
+                    "治_験_の",
+                    "栓_",
+                    "果_を_",
+                    "果_を",
+                    "有_",
+                    "州_",
+                    "岡_",
+                ],
+            ),
+            (
                 "ladino",
                 vec![
                     "a_", "s_", "_d", "e_", "de", "_e", "n_", "en",
@@ -3111,13 +3536,4 @@ impl TextCat {
             ),
         ])
     }
-}
-
-///
-/// Load default categories
-/// We should never pay the price of decoding a JSON which is already compiled.
-///
-/// Future versions will generate code that will not rely on serde for embedded deserialization.
-pub fn languages() -> FileContent {
-    FileContent::new()
 }
