+++ tests/samples.json
@@ -63,5 +63,12 @@
 			"ابوجعفر خازن خراسانی ستاره‌شناس و ریاضی‌دان ایرانی[۱] است که در قرن چهارم هجری قمری در خراسان می‌زیست. از او کتاب‌هایی در این دو علم در دست است. وی به کمک مقاطع مخروطی، معادلهٴ درجه سوم را که به معادلهٴ ماهانی موسوم است حل کرد. در مابین سال‌های ۹۶۱ و ۹۷۱ میلادی درگذشت.",
 			"خراسان (خراسان بزرگ یا خراسان باستانی یا خراسان تاریخی) سرزمینی بزرگ‌تر از خراسان امروزی، بخشی از ایران بزرگ و از دوران ایران ساسانی نامی سنتی و همگانی برای اشاره به نواحی شرقی ایران است. ساسانیان سرزمین ایران را چهار بخش کرده بودند که یکی از آن بخش‌ها خراسان به معنای «سرزمین خورشید» بود.[۲][۳] گسترهٔ تاریخی خراسان بزرگ استان خراسان در ایران کنونی و بخش‌های شمالی و باختری افغانستان و ترکمنستان امروزی را در بر می‌گرفت. در مفهوم گسترده‌تر ازبکستان و تاجیکستان و بخش‌هایی از قرقیزستان و قزاقستان را هم می‌توان در خراسان بزرگ تاریخی برشمرد.[۴]"
 		]
+	},
+	{
+		"category": "japanese",
+		"samples": [
+			"呉座氏は、ベストセラー「応仁の乱」で知られる日本中世史が専門の歴史学者。「鎌倉殿の13人」は三谷幸喜さんが脚本を担当し、鎌倉幕府の2代執権北条義時を描く。主演の義時を小栗旬さんが演じる",
+			"などと批判の声が上がっていた"
+		]
 	}
 ]
